Rework venue addition logic
Closes #4. Previously, venues currently under consideration were stored
in a session variable and an integer index was passed from the frontend
to indicate which venue to add. This approach is flawed, not only
because it made an AJAX request every time the user adds/removes a venue
into the temporary list, but also it limited the choice of venues to the
set currently in memory.

The new approach puts all the information needed to add venues in the
request and does not rely on information stored in the session.